autonomous-ai-assistant/requirements.txt
# AI & NLP Frameworks
transformers==4.35.2
spacy==3.7.2
huggingface-hub==0.20.3
tqdm
scikit-learn
scipy
nltk
Pillow==10.1.0
# Vector Database
faiss-cpu==1.7.4
numpy==1.26.2

# Web Server for API
flask==3.0.0

# Language model for spaCy
# en_core_web_sm